Visual mode, UDMF: added rendering support for fogdensity and floor/ceiling glow properties. Internal, build tools: use origin/master to get commits count and current hash instead of local master.

+++ Source/Core/GZBuilder/Data/GlowingFlatData.cs
@@ -5,7 +5,7 @@ namespace CodeImp.DoomBuilder.GZBuilder.Data
 	public class GlowingFlatData
 	{
 		public PixelColor Color;
-		public int Height;
+		public float Height;
 		public int Brightness = 255;
 		public bool Fullbright;
 		public bool CalculateTextureColor;
